+++ test/src/analyzers/unused_l10n_analyzer/unused_l10n_analyzer_test.dart
@@ -14,7 +14,6 @@ void main() {
       const rootDirectory = '';
       const analyzerExcludes = <String>[
         'test/resources/**',
-        'test/resources/unused_files_analyzer/generated/**/**',
         'test/**/examples/**',
       ];
       final folders = [